[RMD] /W4

+++ DrawPage.h
@@ -135,7 +135,7 @@ private:
 struct CRozm final
 {
     CRozm() noexcept : CRozm(468, 575, 40, 34, 5, 6, 0, false) {}
-    CRozm(int m_w, int m_h, int m_sw, int m_sh, BYTE m_szpalt_x, BYTE m_szpalt_y, int m_typ_xx, bool m_scale_it) noexcept :
+    CRozm(WORD m_w, WORD m_h, WORD m_sw, WORD m_sh, BYTE m_szpalt_x, BYTE m_szpalt_y, int m_typ_xx, bool m_scale_it) noexcept :
         w(m_w),
         h(m_h),
         sw(m_sw),
